+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+
+- When parsing files from the include path, e.g. when looking up external
+  symbols, browse for the corresponding configuration file to correctly set up
+  the environment.
+
 ## [v0.2.0] - 2020-08-24
 
 ### Added